delivery_label/requirements.txt
# 发货标签模块依赖
Pillow>=8.0.0
qrcode[pil]>=7.0.0
